requirements.txt
aiohttp
aiosignal
alembic
anyio
apache-airflow
apache-airflow-providers-common-sql
apache-airflow-providers-ftp
apache-airflow-providers-http
apache-airflow-providers-imap
apache-airflow-providers-sqlite
apispec
argcomplete
asgiref
async-timeout
attrs
Babel
backoff
blinker
cachelib
cassandra-driver
cattrs
certifi
cffi
charset-normalizer
click
clickclick
colorama
colorlog
ConfigUpdater
connexion
cron-descriptor
croniter
cryptography
Deprecated
dill
dnspython
docutils
email-validator
Flask
Flask-AppBuilder
Flask-Babel
Flask-Caching
Flask-JWT-Extended
Flask-Limiter
Flask-Login
Flask-Session
Flask-SQLAlchemy
Flask-WTF
frozenlist
geomet
google-re2
googleapis-common-protos
graphviz
grpcio
gunicorn
h11
httpcore
httpx
idna
importlib-resources
inflection
itsdangerous
Jinja2
jsonschema
jsonschema-specifications
kafka-python
lazy-object-proxy
limits
linkify-it-py
lockfile
Mako
Markdown
markdown-it-py
MarkupSafe
marshmallow
marshmallow-oneofschema
marshmallow-sqlalchemy
mdit-py-plugins
mdurl
multidict
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
ordered-set
packaging
pathspec
pendulum
pluggy
prison
protobuf
psutil
py4j
pycparser
pydantic
Pygments
PyJWT
python-daemon
python-dateutil
python-nvd3
python-slugify
pytz
pytzdata
PyYAML
referencing
requests
requests-toolbelt
rfc3339-validator
rich
rich-argparse
rpds-py
setproctitle
six
sniffio
spark
SQLAlchemy
SQLAlchemy-JSONField
SQLAlchemy-Utils
sqlparse
tabulate
tenacity
termcolor
text-unidecode
typing_extensions
uc-micro-py
unicodecsv
urllib3
Werkzeug
wrapt
WTForms
yarl
